4741-LangChain-e-Python-criando-ferramentas-com-a-LLM-OpenAI/requirements.txt
langchain-google-genai
langchain==0.3.25
langgraph==0.4.8
python-dotenv==1.1.0
faiss-cpu==1.8.0
langchain-community==0.3.25
pypdf==4.2.0
langchain-ollama